monolith_alt/requirements.txt
Django==1.11.2
celery==4.1.0
django_constance==2.0.0
django_import_export==0.5.1
tablib==0.12.0
django_autocomplete_light==3.2.9
django-widget-tweaks==1.4.1
django-picklefield==1.0.0
djangorestframework==3.6.4
pika==0.11.0
Pillow
unicode_slugify==0.1.3

